fix: reduce memory usage in archive file ingestion  (#480)
Previously, when parsing and archive tsv to determine the canonical
fork, we would always start at block 0, even if there was a rocksdb
instance containing stacks chaindata. The entire fork was loaded in
memory and inserted.

Now, we start at the latest confirmed chaintip for this process, which
greatly reduces the memory footprint on subsequent startups of the
Chainhook service. Note: an initial run of the Chainhook service can
still use significant memory.